refactor audit stuff out of index
