+++ src/index.js
@@ -2,20 +2,19 @@ import * as R from "ramda";
 import * as dotenv from "dotenv";
 
 import { App, MemoryStore } from "@slack/bolt";
-import {
-  ACTIONS,
-  COMMANDS,
-  EVENTS,
-  MESSAGES,
-  SETTING_NAMES
-} from "./constants";
+import { ACTIONS, COMMANDS, EVENTS, MESSAGES } from "./constants";
 
 import projectMessage from "./messages/project_message";
 import availableProjects from "./messages/available_projects";
 import projectModal from "./messages/project_modal";
 import itemModal from "./messages/item_modal";
 import sectionModal from "./messages/section_modal";
-import auditPost from "./messages/audit_post";
+import {
+  dumpProjects,
+  postAuditMessageMaker,
+  removeAuditChannel,
+  setAuditChannel
+} from "./auditing";
 
 import {
   addItem,
@@ -24,20 +23,16 @@ import {
   deleteItem,
   deleteProject,
   deleteSection,
-  deleteSetting,
   getFullProject,
   getItemById,
   getProjectById,
   getProjects,
-  getProjectsDump,
   getSectionById,
   moveItem,
   moveSection,
   updateItem,
   updateProject,
-  updateSection,
-  getSetting,
-  setSetting
+  updateSection
 } from "./models";
 
 dotenv.config();
@@ -56,16 +51,7 @@ const convoStore = new MemoryStore();
   console.log(`Bolt app is running on port ${port}`);
 })();
 
-const postAuditMessage = async (userId, projectName, changeText, token) => {
-  const auditChannelSetting = await getSetting(SETTING_NAMES.auditChannelId);
-  if (auditChannelSetting) {
-    app.client.chat.postMessage({
-      token,
-      channel: auditChannelSetting.value,
-      blocks: auditPost(userId, projectName, changeText)
-    });
-  }
-};
+const postAuditMessage = postAuditMessageMaker(app);
 
 const lookupProject = async (projectName, editable, respond, token) => {
   const project = await getFullProject(projectName);
@@ -84,60 +70,6 @@ const lookupProject = async (projectName, editable, respond, token) => {
       blocks: projectMessage(project, editable)
     });
   }
-};
-
-const dumpProjects = async (respond, token) => {
-  const projects = await getProjectsDump();
-  R.map(project => {
-    return respond({
-      token,
-      replace_original: false,
-      response_type: "in_channel",
-      blocks: projectMessage(project, false, true)
-    });
-  }, projects);
-};
-
-const setAuditChannel = async (respond, token, channelId, channelName) => {
-  if (channelName === "directmessage") {
-    return respond({
-      token,
-      response_type: "ephemeral",
-      text: MESSAGES.auditChannel.notDM()
-    });
-  }
-  const currentChannelId = await getSetting(SETTING_NAMES.auditChannelId);
-  if (currentChannelId && currentChannelId !== channelId) {
-    return respond({
-      token,
-      response_type: "ephemeral",
-      text: MESSAGES.auditChannel.dropFirst()
-    });
-  }
-
-  await setSetting(SETTING_NAMES.auditChannelId, channelId);
-  return respond({
-    token,
-    response_type: "in-channel",
-    text: MESSAGES.auditChannel.set(channelName)
-  });
-};
-
-const removeAuditChannel = async (respond, token, channelId) => {
-  const currentChannelSetting = await getSetting(SETTING_NAMES.auditChannelId);
-  if (currentChannelSetting && currentChannelSetting.value !== channelId) {
-    return respond({
-      token,
-      response_type: "ephemeral",
-      text: MESSAGES.auditChannel.dropWrongChannel()
-    });
-  }
-  await deleteSetting(SETTING_NAMES.auditChannelId);
-  return respond({
-    token,
-    response_type: "in-channel",
-    text: MESSAGES.auditChannel.dropped()
-  });
 };
 
 app.command("/alfred", async ({ command, ack, respond, context, body }) => {
